pd02

+++ jupyter/nb/source/pd02_cuda/include/cuda_race.cu
@@ -4,7 +4,7 @@
 
 #include "cuda_util.h"
 
-__global__ void cuda_thread_fun(unsigned long * p, int n) {
+__global__ void cuda_thread_fun(unsigned long long * p, int n) {
   int i        = blockDim.x * blockIdx.x + threadIdx.x;
   if (i < n) {
 /*** if VER == 1 or VER == 2 */
@@ -20,12 +20,12 @@ int main(int argc, char ** argv) {
   int thread_block_sz = (argc > 2 ? atoi(argv[2]) : 64);
   int n_thread_blocks = (n + thread_block_sz - 1) / thread_block_sz;
 
-  unsigned long c;
-  unsigned long * c_dev;
-  check_api_error(cudaMalloc(&c_dev, sizeof(unsigned long)));
+  unsigned long long c;
+  unsigned long long * c_dev;
+  check_api_error(cudaMalloc(&c_dev, sizeof(unsigned long long)));
   check_launch_error((cuda_thread_fun<<<n_thread_blocks,thread_block_sz>>>(c_dev, n)));
   check_api_error(cudaDeviceSynchronize());
-  check_api_error(cudaMemcpy(&c, c_dev, sizeof(unsigned long), cudaMemcpyDeviceToHost));
+  check_api_error(cudaMemcpy(&c, c_dev, sizeof(unsigned long long), cudaMemcpyDeviceToHost));
   check_api_error(cudaFree(c_dev));
   printf("c = %lu\n", c);
   return 0;
